feat: extract exit rule helpers
